Fix Popbill method safety guard
